+++ zcash_primitives/src/transaction/builder.rs
@@ -53,8 +53,8 @@ use crate::{
 #[cfg(zcash_unstable = "nu6")] /* TODO nu7 */
 use crate::transaction::builder::Error::{IssuanceBuilder, IssuanceBundle};
 use orchard::note::AssetBase;
-use orchard::orchard_flavors::OrchardVanilla;
-#[cfg(zcash_unstable = "nu6")] /* TODO nu7 */ use orchard::orchard_flavors::OrchardZSA;
+use orchard::orchard_flavor::OrchardVanilla;
+#[cfg(zcash_unstable = "nu6")] /* TODO nu7 */ use orchard::orchard_flavor::OrchardZSA;
 #[cfg(zcash_unstable = "nu6")] /* TODO nu7 */
 use orchard::{
     issuance::{IssueBundle, IssueInfo},
